Refactor Flask app into submodules
