+++ retrorecon/routes/notes.py
@@ -1,21 +1,31 @@
 import app
 from flask import Blueprint, request, jsonify
+from retrorecon.services import (
+    load_saved_tags,
+    save_saved_tags,
+    get_notes,
+    add_note,
+    update_note,
+    delete_note_entry,
+    delete_all_notes,
+    export_notes_data,
+)
 
 bp = Blueprint('notes', __name__)
 
 @bp.route('/saved_tags', methods=['GET', 'POST'])
 def saved_tags_route():
     if request.method == 'GET':
-        return jsonify({'tags': app.load_saved_tags()})
+        return jsonify({'tags': load_saved_tags(app.SAVED_TAGS_FILE)})
     tag = request.form.get('tag', '').strip()
     if not tag:
         return ('', 400)
     if not tag.startswith('#'):
         tag = '#' + tag
-    tags = app.load_saved_tags()
+    tags = load_saved_tags(app.SAVED_TAGS_FILE)
     if tag not in tags:
         tags.append(tag)
-        app.save_saved_tags(tags)
+        save_saved_tags(app.SAVED_TAGS_FILE, tags)
     return ('', 204)
 
 @bp.route('/delete_saved_tag', methods=['POST'])
@@ -25,17 +35,17 @@ def delete_saved_tag():
         return ('', 400)
     if not tag.startswith('#'):
         tag = '#' + tag
-    tags = app.load_saved_tags()
+    tags = load_saved_tags(app.SAVED_TAGS_FILE)
     if tag in tags:
         tags.remove(tag)
-        app.save_saved_tags(tags)
+        save_saved_tags(app.SAVED_TAGS_FILE, tags)
     return ('', 204)
 
 @bp.route('/notes/<int:url_id>', methods=['GET'])
 def notes_get(url_id: int):
     if not app._db_loaded():
         return jsonify([])
-    rows = app.get_notes(url_id)
+    rows = get_notes(url_id)
     return jsonify([
         {
             'id': r['id'],
@@ -57,9 +67,9 @@ def notes_post():
         return ('', 400)
     note_id = request.form.get('note_id', type=int)
     if note_id:
-        app.update_note(note_id, content)
+        update_note(note_id, content)
     else:
-        app.add_note(url_id, content)
+        add_note(url_id, content)
     return ('', 204)
 
 @bp.route('/delete_note', methods=['POST'])
@@ -68,9 +78,9 @@ def delete_note_route():
     url_id = request.form.get('url_id', type=int)
     delete_all = request.form.get('all', '0') == '1'
     if note_id:
-        app.delete_note_entry(note_id)
+        delete_note_entry(note_id)
     elif url_id and delete_all:
-        app.delete_all_notes(url_id)
+        delete_all_notes(url_id)
     else:
         return ('', 400)
     return ('', 204)
@@ -79,5 +89,5 @@ def delete_note_route():
 def export_notes():
     if not app._db_loaded():
         return jsonify([])
-    data = app.export_notes_data()
+    data = export_notes_data()
     return jsonify(data)
