+++ retrorecon/routes/settings.py
@@ -86,4 +86,4 @@ def set_items_per_page():
     if count not in app.ITEMS_PER_PAGE_OPTIONS:
         return ('Invalid value', 400)
     session['items_per_page'] = count
-    return redirect(url_for('index'))
+    return redirect(url_for('urls.index'))
